[RF] use constexpr BracketAdapter for all scenarios in RooBatchCompute.
New templated code to generate functions that use the constexpr BracketAdapter for all possible combinations of scalar and vector arguments. This enables vectorisation also when agruments other than the first one are vectors. This change makes the libRooBatchCompute grow from 250K to 4.3M.

The roofit/batchcompute/CMakeLists.txt is updated to ignore unused variables because this is unaviodable in the templated code.