[RF] Introduce RooFitDriver and RooNLLVarNew classes
To further develop the batched evaluation mode and run it on the GPU, it
is very inconvenient to use the existing test statistic functions such
as RooNLLVar.

This commit introduces two more classes inheriting from RooAbsReal that
aim to replace the RooNLLVar for the batched evaluations mode.

The RooNLLVarNew class represents the actual computation of the NLL from
a pdf.

The RooFitDriver class manages a RunContext object to evaluate a
RooAbsReal in batched mode and serves as an interface to the
RooMinimizer, which expects a RooAbsReal that can be evaluated in scalar
mode.