[RF] Refactor RooFit pythonization code to reflect inheritance hierachy
The unit test `rooabsreal_ploton` does not work. The problem is that the
`RooAbsReal::plotOn` pythonization only applies to `RooAbsReal`, but it
did not apply to the override in `RooAbsPdf`.

To solve this problem, the pythinization functions are now organized in
classes that reflect the original inheritance hierachy in RooFit.