[RF] Analytical integration for RooNDKeysPdf with custom range
Custom ranges seemed to be supported for analytical integration in
RooNDKeysPdf, but the code was not activated. Numerical integration was
always used for custom ranges.

Unfortunately, using the existing code for analytical integration with
custom ranges gives the wrong result. This commit suggests how this can
be done with new and simpler code that has a logic similar to
RooNDKeysPdf::evaluate() to get the correct result.