[RF] Add UniqueId to RooAbsData
It seems to be important to uniquely identify RooDataSet objects,
otherwise there would be no memory pool mechanism for RooDataSet that
ensures the memory address for each RooDataSet allocated on the heap is
unique.

A `_uniqueId` class member is added to RooAbsData, which should be used
instead of memory address comparisons.