[RF] Remove bin adding functions of `RooHist` from public interface
The functions to add bins to a `RooHist` were removed from the public
interface, because the RooHist is an implementation detail of RooFit
plotting:

  * `addBin`, `addBinWithError`, `addBinWithXYError`, `addAsymmetryBin`,
    `addAsymmetryBinWithError`, `addEfficiencyBin`, and
    `addEfficiencyBinWithError`

Having them in the public interface increases the risk of creating a
`RooHist` with inconsistent bin width corrections or scale factors.

A constructor to create a RooHist from content arrays was introduced
instead.

It also allows us to change the behavior of these functions without
causing unexpected behavior changes in user code.

As a corrolary, the non-IO constructor to create an empty RooHist was
also removed.