[RF] Make RooCacheManager not a templated class
The `RooCacheManager` was a templated class so far, but the only
instance was `RooCacheManager<RooAbsCacheElement>` because
RooObjCacheManager inherits from that.

This is an awkward mixture of template metaprogramming and object
orientation. Both the templating and the `RooAbsCacheElement` achieved
the same thing: allowing for different classes to be cached.

The object oriented approach is used all the way, as the
`RooCacheManager` is never used outside of the `RooObjCacheManager`,
meaning the templating is unnecessary in the RooFit implementation
details. The templating is actually a nuisance: if one changes the
implementation of the RooCacheManager or the `RooNormSetCache` used
within it, all of RooFit needs to be recompiled. This is a nuisance in
debugging, especially on slower machines like the CI build nodes.

Therefore, the RooCacheManager is now changed to always use
`RooAbsCacheElement` instead of a template argument.

In the same go, the dictionary generation for the RooCacheManager is
removed, because caching classes never take part in IO.

The code that got moved from header files to `cxx` files also got ironed
over with clang-format.