[RF] Avoid dummy integral object for already self-normalized RooPoisson
The RooPoisson is implemented with TMath::Poisson, which is normalized
if x covers the full definition range of the Poisson distribution, which
is zero to infinity. By correctly reporting this self-normalization in
that case, the creation of a dummy integral object that returns one is
avoided. The reduced overhead results in a significant speedup of
HistFactory fits with gamma constraints between 10 and 15 % in the hf001
tutorial example.