[RF] Validate alternative bin boundaries in `RooRealVar::getBinning()`
An alternative binning where the bounds are outside the bounds of the
default binning is meaningless, because a variable can't be set to a
value outside the default bounds.

This should be prevented in the RooFit design. A quick check needs to be
made everytime an alternative binning is retrieved, because the default
bounds can also change after the binning is constructed.

With this check, one also doesn't need to worry anymore that bin centers
are outside the variable definition range in other code paths.