[RF] Slim down interface of `getWeightBatch` of dataset classes
The `getWeightBatch` function had some parameters to request the weights
only for a subset of entries. This is not necessary, because the caller
can always get a subset of the returned span also with minimal overhead.

Anyway, this is almost done nowhere in RooFit except for the old
BatchMode implementation. To declutter the interface of this method that
was only introduced in 6.26 anyway (so it can still be changed), this
commit suggests to remove the subrange parameteres and rename the
function to the clearer `RooAbsData::allWeights()`.