[RF] Remove external weight arrays from data stores
The external weight arrays in the `RooVectorDataStore` and
`RooTreeDataStore` were only used to provide the correct `weight()` and
`weightError()` implementation in the RooDataHist case.

However, the RooDataHist oderloads `weight()` and `weightError()`
directly, so duplicating the implementation in the data stores is
redundant.

This commit suggests to remove these external weight arrays from the
datastores and get git of the histogram-specific error calculations.
This avoids code duplication, avoids boilerplate code in the
RooDataHist, and makes it more obvious what the data stores are actually
used for.