[RF] Inline infinity checks in `RooNumber` and mark them `constexpr`
This is mainly a performance optimization, because calling a small
funciton that uses a static variable is quite expensive, given the
simple infinity check.

Also, it is not necessary to have a `ClassDef` macro for RooNumber,
which is only a simple helper class to deal with infinities.

Furthermore, the code hidden behind the `HAS_NUMERIC_LIMITS`
preprocessor define is removed, because this was never used. The macro
is never defined in any build configuration.