[RF] Make the `RooAbsRealLValue::inRange()` overloads consistent
The different overloads of `RooAbsRealLValue::inRange()` implemented
different tolerances when checking if a value `x` falls inside a
specific range. Some overloads checked if the interval
`[x - 1e-6, x + 1e6]` is overlapping with the range, an other overload
checked if the interval `[x - 1e-8*x, x + 1e8*x]` is overlapping.

It's better is this is done consistently, so this commit suggests to do
the more sensible of the two checks (to use a relative epsilon) for all
overloads.