[RF] Register correct servers in RooTruthModel
The `RooTruthModel` when compiled for a basis registered only the
`RooFormulaVar` for the basis as a value server, instead of replicating
the server structore of the `RooFormulaVar` as done in the general
`RooResolutionModel`.

While this was the correct thing to do for the "generic basis" case
where the RooFormulaVar was actually evaluated, it was wrong for the
case where it is not evaluated (any other case).

This commit changes the `RooTruthModel::changeBasis()` function such
that in the non-generic basis case, it will take over the servers of the
`RooFormulaVar`, just like in the implementation of the
`RooResolutionModel` base class.

This change is done to avoid unnecessary evaluation and copying in the
RooFit GPU BatchMode.

Also, a memory leak is fixed where the old basis was not deleted in case
the resolution model owned the basis.