[RF] Restrict to positive decay times in `rf704_amplitudefit` tutorial
Negative decay times are not physical, and therefore the domain of the
time observable should only start at zero.

The `stressRooFit` reference file needs to be adjusted to accomodate for
that.

This is changed now because the tutorial uses the RooTruthModel, which
always returns zero if the observable is not positive and it is
convoluted with a basis that is only defined in the positive domain.

So far, the plot in the output of the tutorial extended to the negative
domain anyway, because there was a typo in the `RooFormulaVar` passed as
a basis function: it contained a whitespace. Therefore, the
RooTruthModel was not actually invoked, but it simply forwarded the
value of the basis RooFormulaVar.

This typo was only an extra whitespace. I will suggest to ignore whitespaces
difference in the basis formula matching for the analytical convolution
in the next commit, and with the correct matching the decay model would
be zero at negative times. That means the reference results would have
to be updated anyway, even if the domain of the time observable would
not be changed.