[RF] Avoid expensive calls to printStream in RooAbsReal::logEvalError
If one doesn't request printing the evaluation errors, the expensive
call to `printStream(oss2,kName|kClassName|kArgs,kInline)` should not be
done. It's result would only be used to set
`_evalErrorList[this].first`, which is not even used anywhere.
This commit suggests to just fill it with the name to get rid of this
overhead when having evaluation errors. This reduces the fitting times
in the `testNaNPacker` tests by a factor two.