[RF] Move `RooStats::ModelConfig` to RooFitCore
The `RooStats::ModelConfig` is suggested to ship with the `RooFitCore`
packages because it makes more sense from a dependency perspective.

The only thing that is usually used from RooStats is the ModelConfig,
and it actually takes a central role in adding information to a
RooWorkspace. Therefore, it makes more sense to put the ModelConfig into
RooFitCore such that the I/O of workspaces doesn't introduce RooStats as
a dependency.

For backwards compatibility, the ModelConfig is still in the RooStats
namespace, but there is now an alias `RooFit::ModelConfig`.