[RF] Improve limit finding code in RooStats LikelihoodInterval
Suggests to use more idiomatic RooFit code in RooStats.

Also, don't check the Minos output code, because if the Minos errors are
at the limit, Minuit2 will return nonzero, unlike Minuit 1.
This case where the errors are at limit should not cause a failure
though. The `stressRooStats` test suite actively provokes it, by fitting
a dataset with a single observed value that corresponds to a best-fit
parameter value at the limit.

With these changes, the `stressRooStats` suite will also pass with
Minuit2 and BatchMode, which is now added as a unit test.