[hist] Made TH1::CheckConsistency public
Made the static function `TH1::CheckConsistency()` and its return type
public.

At the moment, `TH1::CheckConsistency` is called internally by
`TH1::Add`, and prints error messages in case of failure, but continues
the execution. Making it public allows user code to manually check the
consistency to catch errors and stop the program. This is especially
useful for programs that make hundreds of `Add`s, out of which only a
few fail, and the error messages may be lost in logs and go unnoticed.