[RF] Explicitly use legacy backend for references in `testRooRealL`
The new multiprocessing test statistics were designed to be bit-by-bit
compatible with the old test statistics, not the new CPU evaluation
backend. Therefore, the old NLL evaluation backend has to be used for
the offsetting tests (the new CPU backend is not bit-by-bit compatible
with the legacy backend).

This is a followup to commit 65b15d333c, which didn't apply this change
to all relevant tests. Doing `EXPECT_EQ` on floats is quite fragile...