[CMake] Skip connection check if `fail-on-missing=ON`
The connection check only makes sense for `fail-on-missing=OFF`, where
the result is used to decide whether to download a missing dependency as
a builtin from the internet, or to disable the feature that has the
missing dependency.

With `fail-on-missing=ON`, it doesn't matter because disabling features
is not allowed. Therefore, we can skip the connection check to save some
configuration overhead and just assume we have internet: if a builtin
can't be downloaded there will be a configuration failure either way.

Closes #11603 without introducing an additional flag.