[RF] Work around ambiguities in RooFitResult comparisons
The next commit will make the way how fit results are filled more
consistent: the list of floating (initial and final) and constant
parameters will only depend on the current state of the parameters,
without hysteresis effects of changing "isConstant" flags during the
lifetime of the RooMinimizer.

To avoid updating the reference results for now, we ignore
the information that was changed in the comparison.