[RF] Take out RooFit code generation context outside of Detail namespace
We are expecting users to interact with this class, since it is
essential to implement code generation for your own RooFit primitives.

Putting the class in the "Detail" namespace sends the wrong message and
normalized it for users to ignore the convention that "Detail" or
"Internal" stuff should not be used.

While removing the namespace, also rename the class to a shorter and
more succinct name, matching the "codegen" evaluation backend name.
"Codegen" also sounds more friendly than "CodeSquash", which reminds me
of "fly squash".