[Python] Fix RDF Pythonization tests with `builtin_llvm=OFF`
On my machine, the RDF pythonization tests don't run when I build ROOT
with `builtin_llvm=OFF`. There is a crash unless I do `import numba` in
the beginning.

I guess it's not a priority to understand the underlying problem because
not many people build ROOT like this and the workaround is easy, but at
least the tests should be green.

We do the same "magically ordered imports" also in other Python tests
where ROOT doesn't work together with xgboost because of `std::regexp`
symbol clashes, unless you do the import in a certain order:

https://github.com/root-project/root/pull/15183