[tmva][sofie] Restructure emitted code to be differentiable with Clad
The idea of this commit is to refactor the `doInfer()` function that
implements the inference from a member function of the `Session` struct
to a free function that takes the `Session` by `const`-reference.

This free function should only use the session struct and bare C-style
arrays, so that Clad will have no problem differentiating it.

A unit test for the differentiation of a simple MLP is implemented,
embedded in the existing SOFIE tests.