[CMake] Support standalone builds of Clang
If Clang was installed standalone (in particular, not in the LLVM tree),
use the CLANG_INSTALL_PREFIX variable to indicate where it is.

This is the same as 02b0dc129d0c62, but only changing the behavior when
`CLANG_INSTALL_PREFIX` is set without any other changes. This should
be a much less risky change (see also the revert in 61797364d3464).