[Math] Define behavior of `TMath::BinarySearch` for underflow values
The function `TMath::BinarySearch` finds the last element in a sorted
collection that is smaller than the looked up value. In the current
implementation, looking up values that are smaller than the smallest
value is undefined behavior, which raises MSVC debug assertion failures.

The fix suggested by @bellenot in #17179 was to just return the iterator
to the first element in that case, but that means the lookup value x is
mathematically "clipped" to the smallest value in the collection. This
is not necessarily useful and can give you false positives.

Then, the alternative suggestion by @lmoneta in [this comment](https://github.com/root-project/root/pull/17179#issuecomment-2553373256)
was to change the return type to integer and return "-1" if the value is
smaller than the first element. But this defeats the purpose of having
an overload for iterators: in the general case where the lookup is not
done in constant time, returning the index to the found element instead
of an iterator is not useful because then you need to lookup the values
again by index. Also, overloads that work with C arrays and return
integers already exist.

The alternative solution that this commit suggests is to return the
`end()` iterator in these cases where the value is smaller than the
first element. So far, this return value is not used, so there is no
overload of semantics, and also it's quite common to return the `end()`
operator if there is no hit. And in this case, there is no element that
is smaller than the value, so it indeed corresponds to "nothing found".

Also, re-implement the function to actually take arbitrary STL
collections, and not restrict it to things like `std::vector` by using
arithmetic operators on iterators.

Alternative to #17179.