[core] Don't add ROOT library path to dynamic library path on Unix
As explained in https://its.cern.ch/jira/browse/ROOT-9569, the
`TROOT::GetLibDir()` is wrong anyway if ROOT was built with
`gnuinstall=OFF` and the `CMAKE_INSTALL_LIBDIR` was not the default
`lib`.

Fortunately, we don't even need to append the ROOT library path, because
we now always use the runpath mechanism to find the ROOT libraries.

Closes https://its.cern.ch/jira/browse/ROOT-9569.

Follows up on 26d24de5d, which makes sure the runpath is always set on
Unix systems.