[rootx] Turn `root` executable into symbolic link to `root.exe`
There is no need to two separate executables anymore, but for backwards
compatibility we keep `root.exe` and then `root` is a symbolic link to
it.

The two things that the `root` executable did was to set `ROOTSYS` and
start the `root.exe` with with full path. But that's not needed, because
`root` and `root.exe` were in the same directory, so if `root` was in
the path already, `root.exe` will be too. The value of `ROOTSYS` can be
inferred from the opened shared libraries, which can always correctly be
opened because the `RPATH` is now always set.